pacify clang 3.2
git-svn-id: https://svn.r-project.org/R/trunk@62627 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/X11/dataentry.c
@@ -2731,7 +2731,7 @@ static int last_wchar_bytes(char *str)
     memset(wcs, 0, sizeof(wcs));
     memset(&mb_st, 0, sizeof(mbstate_t));
 
-    if ((size_t)-1 == (cnt = (int)mbsrtowcs(wcs, (const char **)&mbs, (int)strlen(mbs), &mb_st)))
+    if ((int)-1 == (cnt = (int)mbsrtowcs(wcs, (const char **)&mbs, (int)strlen(mbs), &mb_st)))
     {
         return 0;
     }
